Move AntennaProfile class to EMANE::Events namespace
All other classes that represent an entry in an event (including
Location, Pathloss, CommEffect, SlotInfo, and SlotStructure) are in
the EMANE::Events namespace. Move the AntennaProfile class to this
namespace for consistency.
